+++ NEWS.md
@@ -1,5 +1,8 @@
 
 # News
+# survobj 2.1.0
+Addition of recurrent event simulaion under homogeneous and non homogeneous
+Poisson process
 
 # survobj 2.0.0
 Addition of new distributions and new random generators functions